あああ

+++ 0227version/vq.py
@@ -251,7 +251,7 @@ class ContrastiveLoss(nn.Module):
         latent_similarity_matrix = torch.mm(z, z.T)
 
         if chunk % 100 == 0:
-            hist = torch.histc(latent_similarity_matrix.cpu(), bins=20, min=-1.0, max=1.0)
+            hist = torch.histc(latent_similarity_matrix.cpu().to(torch.float32), bins=20, min=-1.0, max=1.0)
             print(hist)
 
         def calc_repel_loss(v, simi_matrix, chunk):
